V 0.6.5 - Gestion des erreurs de la partie Sums (CRUD)

+++ routes/sums.js
@@ -30,24 +30,48 @@ router.get('/:idext_sum', function(req, res, next) {
     idext_sum = req.params.idext_sum
 
     var getSumsItem = function(retfunc){
-        connection.query('SELECT idext_sum, COALESCE(SUM(value_item * quantity_sum),0) as total_sum, date_sum FROM sum_item_cashReg si JOIN item_cashReg i ON si.id_item = i.idint_item RIGHT JOIN sum_cashReg s ON si.id_sum=s.idint_sum WHERE idext_sum = ? GROUP BY idint_sum', [idext_sum], function(error, results, fields) {
-            
-            if(error) res.send(error);
-            else {
-                connection.query('SELECT name_item, quantity_sum, value_item, COALESCE((value_item * quantity_sum),0) as total_item FROM sum_item_cashReg si JOIN sum_cashReg s ON si.id_sum = s.idint_sum JOIN item_cashReg i ON si.id_item = i.idint_item WHERE idext_sum = ? ', [idext_sum], function(error, resultItem, fields) {
+        if(!Number.isNaN(idext_sum) && idext_sum > 0){
+            connection.query('SELECT idext_sum, COALESCE(SUM(value_item * quantity_sum),0) as total_sum, date_sum FROM sum_item_cashReg si JOIN item_cashReg i ON si.id_item = i.idint_item RIGHT JOIN sum_cashReg s ON si.id_sum=s.idint_sum WHERE idext_sum = ? GROUP BY idint_sum', [idext_sum], function(error, results, fields) {
+                
+                if(error) res.send(error);
+                else {
+                    if(results.length>0){
+                        connection.query('SELECT name_item, quantity_sum, value_item, COALESCE((value_item * quantity_sum),0) as total_item FROM sum_item_cashReg si JOIN sum_cashReg s ON si.id_sum = s.idint_sum JOIN item_cashReg i ON si.id_item = i.idint_item WHERE idext_sum = ? ', [idext_sum], function(error, resultItem, fields) {
 
-                    if(error) res.send(error);
-                    else {
-                        if(resultItem.length>0){
-                            results[0]['items'] = resultItem
-                        }
-                        retfunc(results);
+                            if(error) res.send(error);
+                            else {
+                                if(resultItem.length>0){
+                                    results[0]['items'] = resultItem
+                                }
+                                retfunc(results);
+                            }
+                            
+                        });
                     }
-                    
-                });
-            }
+                    else {                        
+                        var error = {
+                            'error_id' : 13,
+                            'error_type' : "ERR_ID_NON_EXISTENT",
+                            'error_value' : idext_sum,
+                            'error_var' : "idext_sum",
+                            'error_text' : "Erreur ERR_ID_NON_EXISTENT - La valeur de idext_sum ("+idext_sum+") n'existe pas."
+                        }
+                        retfunc(error)
+                    }
+                }
 
-        });
+            });
+        }
+        else {
+            var error = {
+                'error_id' : 11,
+                'error_type' : "ERR_VALUE_FORMAT",
+                'error_value' : idext_sum,
+                'error_var' : "idext_sum",
+                'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_sum ("+idext_sum+") est incorrect."
+            }
+            retfunc(error)
+        }
     }
 
     getSumsItem(function(results) {
@@ -63,24 +87,36 @@ router.post('/', function(req, res, next) {
     idext_sum = req.body.idext_sum;
 
     var postSum = function(retfunc){
-        connection.query('SELECT idext_sum FROM sum_cashReg WHERE idext_sum = ?', [idext_sum], function(error, results_idext, fields) {
-            if(results_idext.length <= 0){
-                if( !Number.isNaN(idext_sum) && idext_sum>0 ){
-                    connection.query('INSERT INTO sum_cashReg (idext_sum) VALUES (?)', [idext_sum], function(error, results, fields) {
-                        
-                        if(error) res.send(error);
-                        else retfunc(results);
-
-                    });
+        if( !Number.isNaN(idext_sum) && idext_sum>0 ){
+            connection.query('SELECT idext_sum FROM sum_cashReg WHERE idext_sum = ?', [idext_sum], function(error, results_idext, fields) {
+                if(results_idext.length <= 0){                    
+                        connection.query('INSERT INTO sum_cashReg (idext_sum) VALUES (?)', [idext_sum], function(error, results, fields) {
+                            if(error) res.send(error);
+                            else retfunc(results);
+                        });
                 }
                 else {
-                    res.send("Erreur de format de valeur.");
+                    var error = {
+                        'error_id' : 12,
+                        'error_type' : "ERR_ID_ALREADY_EXISTS",
+                        'error_value' : idext_sum,
+                        'error_var' : "idext_sum",
+                        'error_text' : "Erreur ERR_ID_ALREADY_EXISTS - L'objet ayant pour id idext_sum ("+idext_sum+") est déjà existant."
+                    }
+                    retfunc(error)
                 }
+            })
+        }
+        else {
+            var error = {
+                'error_id' : 11,
+                'error_type' : "ERR_VALUE_FORMAT",
+                'error_value' : idext_sum,
+                'error_var' : "idext_sum",
+                'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_sum ("+idext_sum+") est incorrect."
             }
-            else {
-                res.send("Erreur - L'id renseigné existe déja.");
-            }
-        })
+            retfunc(error)
+        }
     }
 
     postSum(function(results) {
@@ -98,7 +134,7 @@ router.post('/:idext_sum', function(req, res, next) {
     qty_item = req.body.qty_item;
     
     var postSumItem = function(retfunc){
-        if( !Number.isNaN(idext_sum) && idext_sum>0 && !Number.isNaN(idext_item) && idext_item>0 && !Number.isNaN(qty_item) && qty_item>0 ){
+        if( !Number.isNaN(idext_sum) && idext_sum>0 && typeof idext_item === "number" && idext_item>0 && typeof qty_item === "number" && qty_item>0 ){
             connection.query('SELECT idint_sum FROM sum_cashReg WHERE idext_sum = ?', [idext_sum], function(error, results_idsum, fields) {
                 
                 if(error) res.send(error);
@@ -123,10 +159,24 @@ router.post('/:idext_sum', function(req, res, next) {
                         }
                         else {
                             if(results_idsum.length <= 0){
-                                res.send("L'id d'addition renseigné est invalide");
+                                var error = {
+                                    'error_id' : 13,
+                                    'error_type' : "ERR_ID_NON_EXISTENT",
+                                    'error_value' : idext_sum,
+                                    'error_var' : "idext_sum",
+                                    'error_text' : "Erreur ERR_ID_NON_EXISTENT - La valeur de idext_sum ("+idext_sum+") n'existe pas."
+                                }
+                                retfunc(error)
                             }
                             else {
-                                res.send("L'id d'item renseigné est invalide");
+                                var error = {
+                                    'error_id' : 13,
+                                    'error_type' : "ERR_ID_NON_EXISTENT",
+                                    'error_value' : idext_item,
+                                    'error_var' : "idext_item",
+                                    'error_text' : "Erreur ERR_ID_NON_EXISTENT - La valeur de idext_item ("+idext_item+") n'existe pas."
+                                }
+                                retfunc(error)
                             }
                         }
                     });
@@ -134,7 +184,36 @@ router.post('/:idext_sum', function(req, res, next) {
             });
         }
         else {
-            res.send("Le format d'une valeur est invalide");
+            if(typeof idext_item !== "number" || idext_item<=0){
+                var error = {
+                    'error_id' : 11,
+                    'error_type' : "ERR_VALUE_FORMAT",
+                    'error_value' : idext_item,
+                    'error_var' : "idext_item",
+                    'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_item ("+idext_item+") est incorrect."
+                }
+                retfunc(error)
+            }
+            else if(typeof qty_item !== "number" || qty_item<=0){
+                var error = {
+                    'error_id' : 11,
+                    'error_type' : "ERR_VALUE_FORMAT",
+                    'error_value' : qty_item,
+                    'error_var' : "qty_item",
+                    'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de qty_item ("+qty_item+") est incorrect."
+                }
+                retfunc(error)
+            }
+            else {
+                var error = {
+                    'error_id' : 11,
+                    'error_type' : "ERR_VALUE_FORMAT",
+                    'error_value' : idext_sum,
+                    'error_var' : "idext_sum",
+                    'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_sum ("+idext_sum+") est incorrect."
+                }
+                retfunc(error)
+            }
         }
     }
 
@@ -152,38 +231,57 @@ router.delete('/:idext_sum', function(req, res, next) {
 
 
     var deleteSumItem = function(retfunc){
-        connection.query('SELECT idint_sum FROM sum_cashReg WHERE idext_sum = ?', [idext_sum], function(error, results_idsum, fields) {     
-            if(error) res.send(error);
-            else if(results_idsum.length > 0){
-                if(idext_item){
-                    connection.query('SELECT idint_item FROM item_cashReg WHERE idext_item = ?', [idext_item], function(error, results_iditem, fields) {     
-                        if(error) res.send(error);
-                        else if(results_iditem.length > 0){
-                            connection.query('SELECT quantity_sum FROM sum_item_cashReg WHERE id_item = ? AND id_sum = ?', [results_iditem[0]['idint_item'], results_idsum[0]['idint_sum']], function(error, results_qty, fields) {     
+        if(!Number.isNaN(idext_sum) && idext_sum > 0){
+            connection.query('SELECT idint_sum FROM sum_cashReg WHERE idext_sum = ?', [idext_sum], function(error, results_idsum, fields) {     
+                if(error) res.send(error);
+                else if(results_idsum.length > 0){
+                    if(idext_item){
+                        if(typeof idext_item === "number" && idext_item){
+                            connection.query('SELECT idint_item FROM item_cashReg WHERE idext_item = ?', [idext_item], function(error, results_iditem, fields) {     
                                 if(error) res.send(error);
-                                else if(results_qty.length > 0){
-                                    if(results_qty[0]['quantity_sum'] == 1){
-                                        connection.query('DELETE FROM sum_item_cashReg WHERE id_sum = ? AND id_item = ?', [results_idsum[0]['idint_sum'], results_iditem[0]['idint_item']], function(error, results, fields) {
-                                            if(error) res.send(error);
-                                            else retfunc(results);
-                                        });
-                                    }
-                                    else{
-                                        connection.query('UPDATE sum_item_cashReg SET quantity_sum = quantity_sum - 1 WHERE id_item = ? AND id_sum = ?', [results_iditem[0]['idint_item'], results_idsum[0]['idint_sum']], function(error, results, fields) {
-                                            if(error) res.send(error);
-                                            else retfunc(results);
-                                        });
-                                    }
+                                else if(results_iditem.length > 0){
+                                    connection.query('SELECT quantity_sum FROM sum_item_cashReg WHERE id_item = ? AND id_sum = ?', [results_iditem[0]['idint_item'], results_idsum[0]['idint_sum']], function(error, results_qty, fields) {     
+                                        if(error) res.send(error);
+                                        else if(results_qty.length > 0){
+                                            if(results_qty[0]['quantity_sum'] == 1){
+                                                connection.query('DELETE FROM sum_item_cashReg WHERE id_sum = ? AND id_item = ?', [results_idsum[0]['idint_sum'], results_iditem[0]['idint_item']], function(error, results, fields) {
+                                                    if(error) res.send(error);
+                                                    else retfunc(results);
+                                                });
+                                            }
+                                            else{
+                                                connection.query('UPDATE sum_item_cashReg SET quantity_sum = quantity_sum - 1 WHERE id_item = ? AND id_sum = ?', [results_iditem[0]['idint_item'], results_idsum[0]['idint_sum']], function(error, results, fields) {
+                                                    if(error) res.send(error);
+                                                    else retfunc(results);
+                                                });
+                                            }
+                                        }
+                                    });
                                 }
                                 else {
-                                    res.send("Erreur - Id item innexistant");
+                                    var error = {
+                                        'error_id' : 13,
+                                        'error_type' : "ERR_ID_NON_EXISTENT",
+                                        'error_value' : idext_item,
+                                        'error_var' : "idext_item",
+                                        'error_text' : "Erreur ERR_ID_NON_EXISTENT - La valeur de idext_item ("+idext_item+") n'existe pas."
+                                    }
+                                    retfunc(error)
                                 }
                             });
                         }
-                        else res.send("Erreur - Id innexistant");
-                    });
-                }
-                else {
+                        else {
+                            var error = {
+                                'error_id' : 11,
+                                'error_type' : "ERR_VALUE_FORMAT",
+                                'error_value' : idext_item,
+                                'error_var' : "idext_item",
+                                'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_item ("+idext_item+") est incorrect."
+                            }
+                            retfunc(error)
+                        }
+                    }
+                    else {
                         connection.query('DELETE FROM sum_item_cashReg WHERE id_sum = ? ', [results_idsum[0]['idint_sum']], function(error, results, fields) {
                             if(error) res.send(error);
                             else connection.query('DELETE FROM sum_cashReg WHERE idint_sum = ? ', [results_idsum[0]['idint_sum']], function(error, results, fields) {
@@ -191,10 +289,30 @@ router.delete('/:idext_sum', function(req, res, next) {
                                 else retfunc(results);
                             });
                         });
+                    }
                 }
+                else {
+                    var error = {
+                        'error_id' : 13,
+                        'error_type' : "ERR_ID_NON_EXISTENT",
+                        'error_value' : idext_sum,
+                        'error_var' : "idext_sum",
+                        'error_text' : "Erreur ERR_ID_NON_EXISTENT - La valeur de idext_sum ("+idext_sum+") n'existe pas."
+                    }
+                    retfunc(error)
+                }
+            });
+        }
+        else {
+            var error = {
+                'error_id' : 11,
+                'error_type' : "ERR_VALUE_FORMAT",
+                'error_value' : idext_sum,
+                'error_var' : "idext_sum",
+                'error_text' : "Erreur ERR_VALUE_FORMAT - Le format de la valeur de idext_sum ("+idext_sum+") est incorrect."
             }
-            else res.send("Erreur - Id sum innexistant");
-        });
+            retfunc(error)
+        }
     }
 
     deleteSumItem(function(results) {
